+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Sort fake chunks before loading so nearby chunks load first
 - Fix render distance resetting after restart if above 32 (#109)
 - Fix errors about Starlight in log when Starlight is not installed
+- Fix lag spikes caused by chunk serialization on main thread when crossing chunk borders (#95)
 
 ### 4.0.0
 - Update to Minecraft 1.19